requirements.txt
streamlit==1.32.0
pandas==2.2.0
numpy==1.26.3
plotly==5.18.0
python-docx==1.1.0
PyPDF2==3.0.1
openpyxl==3.1.2
pytesseract==0.3.10
pillow==10.2.0
groq==0.4.1